refactor: RestTemplate 설정 추가 및 리팩터링
