+++ src/main/java/com/atwoz/member/infrastructure/oauth/RestTemplateOAuthConnection.java
@@ -1,0 +1,47 @@
+package com.atwoz.member.infrastructure.oauth;
+
+import com.atwoz.member.application.auth.OAuthConnectionManager;
+import com.atwoz.member.domain.auth.RestTemplateConfig;
+import com.atwoz.member.ui.auth.support.oauth.OAuthProperties.OAuthProvider;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
+import org.springframework.stereotype.Component;
+import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.util.MultiValueMap;
+import org.springframework.web.client.RestTemplate;
+
+@Component
+@RequiredArgsConstructor
+public class RestTemplateOAuthConnection implements OAuthConnectionManager {
+
+    private final RestTemplateConfig restTemplateConfig;
+
+    @Override
+    public String getAccessTokenResponse(final OAuthProvider oAuthProvider, final String code) {
+        RestTemplate restTemplate = restTemplateConfig.restTemplate();
+        HttpHeaders httpHeaders = new HttpHeaders();
+        httpHeaders.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
+
+        MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
+        params.add("grant_type", "authorization_code");
+        params.add("client_id", oAuthProvider.getClientId());
+        params.add("redirect_uri", oAuthProvider.getRedirectUri());
+        params.add("code", code);
+
+        HttpEntity<MultiValueMap<String, String>> requestEntity = new HttpEntity<>(params, httpHeaders);
+        return restTemplate.postForObject(oAuthProvider.getTokenUri(), requestEntity, String.class);
+    }
+
+    @Override
+    public String getMemberInfoResponse(final String accessToken, final String userInfoUrl) {
+        RestTemplate restTemplate = restTemplateConfig.restTemplate();
+        HttpHeaders headers = new HttpHeaders();
+        headers.setBearerAuth(accessToken);
+        headers.setContentType(MediaType.APPLICATION_JSON);
+        HttpEntity<Void> requestEntity = new HttpEntity<>(headers);
+
+        return restTemplate.postForObject(userInfoUrl, requestEntity, String.class);
+    }
+}
